Extract compound-types metadata.
The compound mode in AV1 has a wedge mode. The wedge mode has different types of wedges specified by their wedge index and wedge sign. In addition to this, each wedge index has a different shape depending on the type of block. Thus, the wedge codebook is also extracted on each frame as "wedge_types"